+++ test/opendss.jl
@@ -298,4 +298,39 @@ TESTLOG = getlogger(PowerModels)
         @test sol["status"] == :LocalOptimal
         @test isapprox(sol["objective"], 0.0182769; atol = 1e-4)
     end
+
+    @testset "3-bus balanced pv" begin
+        setlevel!(TESTLOG, "warn")
+        @test_warn(TESTLOG, "Converting PVSystem \"pv1\" into generator with limits determined by OpenDSS property 'kVA'",
+                   TPPMs.parse_file("../test/data/opendss/case3_balanced_pv.dss"))
+        setlevel!(TESTLOG, "error")
+
+        tppm = TPPMs.parse_file("../test/data/opendss/case3_balanced_pv.dss")
+
+        @test length(tppm["gen"]) == 2
+        @test all(tppm["gen"]["2"]["qmin"] .== -tppm["gen"]["2"]["qmax"])
+        @test all(tppm["gen"]["2"]["pmax"] .== tppm["gen"]["2"]["qmax"])
+        @test all(tppm["gen"]["2"]["pmin"].values .== 0.0)
+
+        sol = TPPMs.run_tp_opf(tppm, PMs.ACPPowerModel, ipopt_solver)
+
+        @test sol["status"] == :LocalOptimal
+        @test sum(sol["solution"]["gen"]["1"]["pg"] * sol["solution"]["baseMVA"]) < 0.0
+        @test sum(sol["solution"]["gen"]["1"]["qg"] * sol["solution"]["baseMVA"]) < 0.0
+        @test isapprox(sum(sol["solution"]["gen"]["2"]["pg"] * sol["solution"]["baseMVA"]), 0.018345; atol=1e-4)
+        @test isapprox(sum(sol["solution"]["gen"]["2"]["qg"] * sol["solution"]["baseMVA"]), 0.00919404; atol=1e-4)
+    end
+
+    @testset "3-bus unbalanced single-phase pv" begin
+        tppm = TPPMs.parse_file("../test/data/opendss/case3_unbalanced_1phase-pv.dss")
+        sol = TPPMs.run_tp_opf(tppm, PMs.ACPPowerModel, ipopt_solver)
+
+        @test sol["status"] == :LocalOptimal
+
+        @test isapprox(sum(sol["solution"]["gen"]["1"]["pg"] * sol["solution"]["baseMVA"]), 0.0196116; atol=1e-3)
+        @test isapprox(sum(sol["solution"]["gen"]["1"]["qg"] * sol["solution"]["baseMVA"]), 0.00923107; atol=1e-3)
+
+        @test all(sol["solution"]["gen"]["2"]["pg"][2:3] .== 0.0)
+        @test all(sol["solution"]["gen"]["2"]["qg"][2:3] .== 0.0)
+    end
 end